Moved actions into their own file.
